fix in viaLocalCopy(): check if the created temp file is not null before deleting it

+++ src/org/ogf/saga/proxies/namespace/NSDirectoryWrapper.java
@@ -239,7 +239,9 @@ public class NSDirectoryWrapper extends NSEntryWrapper implements NSDirectory {
             } catch(Throwable ex) {
                 // ignored
             }
-            tmp.delete();
+            if (tmp != null) {
+                tmp.delete();
+            }
         }
     }
